requirements.txt
langchain==0.2.11
langchain-core==0.2.23
langgraph==0.1.11
python-dotenv==1.0.1
langchain_anthropic==0.1.20
langchain-openai==0.1.17
langchain_mistralai==0.1.10
playwright==1.43.0
xmltodict==0.13.0
beautifulsoup4==4.12.3
esprima==4.0.1
cssutils==2.11.0
libsass==0.23.0
openai==1.35.7
cohere==5.5.8
langchain-cohere==0.1.8
chromadb==0.4.21
todoist-api-python==2.1.5
replicate==0.31.0
termcolor==2.4.0